+++ src/components/Project/ProjectComponents.module.scss
@@ -30,5 +30,5 @@
 }
 
 @media (min-width: 1280) {
-  @import './page_1280.module.scss';
+  @import './ProjectComponents_1280.module.scss';
 }
